feat: change API to function based read operations
Previously Pop/Peek/Move returned the items directly, in big batches.
This required that the buckets would still be open and users would have
to call some GC() function to clean up. This was no nice API, and now the
retrieved items need to be fetched from a user-passed functions. If needed
outside of that, the items need to be copied.